+++ covobeta/src/components/TestPage/index.js
@@ -11,11 +11,24 @@ import "./TestPage.css";
 //import AddTrip from "../AddTrip";
 //import CoVoTripsFromContext from "./Tests/CoVoTripsFromContext.js";
 //import CoVoTrips from "./Tests/CoVoTrips";
-import CoVoDynamicMenu from "./Tests/CoVoDynamicMenu";
+//import CoVoDynamicMenu from "./Tests/CoVoDynamicMenu";
+import CoVoPlaceChooser from "../CoVoPlaceChooser";
 
 //Descriptions of props, what it does, and what it affects
 //Beginning of implementation
 class TestPage extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      testBox: ""
+    };
+  }
+  testBoxContentChange = value => {
+    this.setState({ testBox: value });
+  };
+  testOnCoVoPlaceChosen = () => {
+    console.log("Chose a place");
+  };
   render() {
     return (
       <div className="test-page">
@@ -25,10 +38,18 @@ class TestPage extends Component {
         {/*<CoVoTrips />*/}
         {/*<CoVoTripsFromContext />*/}
         {/*<AddTrip />*/}
+        {/*<CoVoDynamicMenu />*/}
+        {/*boxName : name of the textbox
+        ** boxContent : text in the boxName
+        ** boxContentChange(text) : function to call when there is some new text
+        ** onCoVoPlaceChosen(text)*/}
         <div className="test-content">
-          <CoVoDynamicMenu />
-          <CoVoDynamicMenu />
-          <CoVoDynamicMenu />
+          <CoVoPlaceChooser
+            boxName="Test place chooser"
+            boxContent={this.state.testBox}
+            boxContentChange={this.testBoxContentChange}
+            onCoVoPlaceChosen={this.testOnCoVoPlaceChosen}
+          />
         </div>
       </div>
     );